Refactor: Break down MobileTopbar.tsx into components
Refactor MobileTopbar.tsx into smaller, reusable components to improve code readability and maintainability. No functionality changes are included in this commit.
